+++ api/src/main/kotlin/uy/kohesive/injekt/api/Registrar.kt
@@ -1,7 +1,4 @@
 package uy.kohesive.injekt.api
 
-interface InjektRegistrar: InjektRegistry, InjektFactory  {
-    fun importModule(submodule: InjektModule) {
-        submodule.registerWith(this)
-    }
-}
+@Deprecated(replaceWith = ReplaceWith("InjektScope", "uy.kohesive.injekt.api.InjektScope"), level = DeprecationLevel.WARNING, message = "When used with modules, replace InjektRegistry with InjektScope for extension functions such as InjektRegistry.registerInjectables()")
+interface InjektRegistrar: InjektRegistry, InjektFactory  {}